文章模块

+++ weblog-module-admin/src/main/java/com/oiazheng/weblog/admin/service/impl/AdminTagServiceImpl.java
@@ -5,7 +5,9 @@ import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.oiazheng.weblog.admin.model.vo.tag.*;
 import com.oiazheng.weblog.admin.service.AdminTagService;
+import com.oiazheng.weblog.common.domain.dos.ArticleTagRelDO;
 import com.oiazheng.weblog.common.domain.dos.TagDO;
+import com.oiazheng.weblog.common.domain.mapper.ArticleTagRelMapper;
 import com.oiazheng.weblog.common.domain.mapper.TagMapper;
 import com.oiazheng.weblog.common.enums.ResponseCodeEnum;
 import com.oiazheng.weblog.common.exception.BizException;
@@ -31,6 +33,9 @@ public class AdminTagServiceImpl extends ServiceImpl<TagMapper, TagDO> implement
 
     @Autowired
     private TagMapper tagMapper;
+
+    @Autowired
+    private ArticleTagRelMapper articleTagRelMapper;
 
     /**
      * 添加标签集合
@@ -101,13 +106,21 @@ public class AdminTagServiceImpl extends ServiceImpl<TagMapper, TagDO> implement
 
     @Override
     public Response deleteTag(DeleteTagReqVO deleteTagReqVO) {
-        // 标签ID
+        // 标签 ID
         Long tagId = deleteTagReqVO.getId();
 
-        // 删除标签
+        // 校验该标签下是否有关联的文章，若有，则不允许删除，提示用户需要先删除标签下的文章
+        ArticleTagRelDO articleTagRelDO = articleTagRelMapper.selectOneByTagId(tagId);
+
+        if (Objects.nonNull(articleTagRelDO)) {
+            log.warn("==> 此标签下包含文章，无法删除，tagId: {}", tagId);
+            throw new BizException(ResponseCodeEnum.TAG_CAN_NOT_DELETE);
+        }
+
+        // 根据标签 ID 删除
         int count = tagMapper.deleteById(tagId);
 
-        return count==1 ? Response.success() : Response.fail(ResponseCodeEnum.TAG_NOT_EXIST);
+        return count == 1 ? Response.success() : Response.fail(ResponseCodeEnum.TAG_NOT_EXISTED);
     }
 
     @Override
